requirements.txt
pytest>=8.0
pytest-cov>=5.0
pytest-playwright>=0.5.0
playwright==1.57.0
ruff==0.14.10
Flask>=3.0
Flask-Login>=0.6.3
Flask-WTF>=1.2.1
PyMySQL>=1.1.0
mysql-connector-python>=8.0.0
python-dotenv>=1.0.0
DBUtils>=3.0.0
cryptography>=46.0.0
argon2-cffi>=23.1.0
Pillow>=11.0.0